backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
sentence-transformers
faiss-cpu
llama-cpp-python
requests